refactor: move `make_kw` to test_interface.py
